Frontend

+++ rm.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 
-sudo docker stop csofpn
-sudo docker rm csofpn
+if [ "$(docker ps -a -f name=csofpn)" ]; then
+    sudo docker stop csofpn
+    sudo docker rm csofpn
+    
+fi
+
+
